Fixed HW2 to include a new method called HW2_Main that runs the application.
The application is ran in the HW2 main method. The second file is simply where the code is stored as instructed in the assignment. (At first I had both methods in the same file which I since realized was incorrect) I fixed the assignment to reflect two Seperate files for HW2.