Updated Activity 10
Extracts Bug Number, Status, Heat, and Title for the first page of the bug tracking page.
Creates a database table named "BugData" (you can modify the table name as needed) with columns for Bug Number, Status, Heat, and Title.
Inserts all the extracted bug information into the database after printing it for the user to see (I added the print out to assist locating the data and left it in to ensure I knew the way to print it first and update to database after to ensure data correctness. (It is a good practice to print what you are web scraping before instantly pushing to database in case there are errors, we do not want to use up storage for no reason. The only table created in this case is the one we want. BINGO