We have added three outputs for activity 15.
I provided two csv files. One picture of the printed output as well. I believe the only "issue" that I may be having is that I had a little bit of a tough time trying to figure out which data was supposed to be extracted and what was not. (I might have not done one of them exactly right) but I have demonstrated the working functionality. ( I did it right, the only thing that would be changed would be the file names in which I am extracting the data.